+++ src/components/pokemonSelectRow.js
@@ -1,4 +1,5 @@
 const { StringSelectMenuBuilder, ActionRowBuilder } = require("discord.js");
+const { pokemonConfig } = require("../config/pokemonConfig");
 
 const buildPokemonSelectRow = (pokemons, data, eventName) => {
     const menuId = {
@@ -10,9 +11,11 @@ const buildPokemonSelectRow = (pokemons, data, eventName) => {
         .setCustomId(`${JSON.stringify(menuId)}`)
         .setPlaceholder('Select a pokemon')
         .addOptions(pokemons.map(pokemon => {
+            const speciesData = pokemonConfig[pokemon.speciesId];
             return {
                 label: `${pokemon.name} (${pokemon._id})`,
                 value: `${pokemon._id}`,
+                emoji: speciesData.emoji,
             }
         }));